# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic API"
  description: "This API is used to create, update, and delete information for a mechanic shop"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: |
      Token Required!
      Please enter your JWT token **with `Bearer` prefix** like:
      ```
      Bearer "JWT TOKEN"
      ```
paths:
...
  /customers/{id}/cars:
    post:
      tags:
        - Customers
      summary: "Endpoint for customer to add a car."
      description: "Endpoint for customer to add a car. Requires token authentication, customer must be logged in to add a car to their account"
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: "This route validates that the customer_id in the payload matches the authenticated token’s sub field"
        - in: body
          name: body
          description: "Information needed to add a car as a customer"
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerCarPayload"
      responses:
        200:
          description: "Car added successfully"
          schema:
            $ref: "#/definitions/CreateCustomerCarPayload"
          examples:
            application/json:
              color: "black"
              customer_id: 4
              id: 2
              make: "BMW"
              model: "M4"
              model_year: 2025
        400:
          description: "Validation Error missing information to create a car"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: Forbidden - Not allowed to perform this action
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Customer not found
          schema:
            $ref: "#/definitions/ErrorResponse"

    get:
      tags:
        - Customers
      summary: "Get all cars for a specific customer"
      description: "Returns a list of all cars associated with the given customer ID"
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: "the ID of the customer whose cars are being retrieved"
      responses:
        200:
          description: "Successfully retrieved cars for the customer"
          schema:
            type: array
            items:
              $ref: "#/definitions/CreateCustomerCarPayload"
        404:
          description: Customer not found
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /cars:
    post:
      tags:
        - Cars
      summary: "Endpoint to create a car for a customer"
      description: "Endpoint to create a car and assign to a customer with details on what type of car it is."
      consumes:
        - application/json
      produces:
        - application/json
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          description: "Car details to be created"
          required: true
          schema:
            $ref: "#/definitions/CreateCarPayload"
      responses:
        201:
          description: "Successfully created a car"
          schema:
            type: object
            properties:
              message:
                type: string
              car:
                $ref: "#/definitions/CreateCarResponse"
          examples:
            application/json:
              message: "Car added"
              car:
                id: 5
                color: red
                customer_id: 4
                make: Ferrari
                model: SF90 Stradale
                model_year: 2025
        401:
          description: Unauthorized - Token missing or invalid
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: Forbidden - Admin access required
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    get:
      tags:
        - Cars
      summary: "Gets all cars in database"
      description: "Endpoint to retrieve all cars in the database"
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: page
          in: query
          type: integer
          required: false
          description: "Page number (default: 1)"
        - name: per_page
          in: query
          type: integer
          required: false
          description: "Number of results per page (default: 10)"
      responses:
        200:
          description: "Retrieved all Cars successfully"
          schema:
            $ref: "#/definitions/AllCars"
          examples:
            application/json:
              cars:
                - id: 1
                  color: black
                  make: BMW
                  model: M4
                  model_year: 2025
                  customer_id: 1
                - id: 2
                  color: blue
                  make: BMW
                  model: M4
                  model_year: 2025
                  customer_id: 1
              page: 1
              per_page: 10
              total_cars: 2
  /cars/{id}:
    get:
      tags:
        - Cars
      summary: "Retrieve a single car by ID"
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: "ID of the car to retrieve"
      responses:
        200:
          description: "Car retrieved successfully"
          schema:
            $ref: "#/definitions/CreateCarResponse"
        404:
          description: Car not found
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      tags:
        - Cars
      summary: "Endpoint to update a Car"
      description: "Endpoint to update a Car, this is a token authenticated route, and the user must be logged in to update"
      consumes:
        - application/json
      produces:
        - application/json
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: "ID of the Car to update"
        - in: body
          name: body
          description: "Information required to update a Car"
          required: true
          schema:
            $ref: "#/definitions/UpdateCarPayload"
      responses:
        200:
          description: "Successfully Updated Car"
          schema:
            $ref: "#/definitions/UpdateCarResponse"
          examples:
            application/json:
              id: 1
              color: black
              make: Toyota
              model: Camry
              model_year: 2025
        401:
          description: Unauthorized - Missing or Invalid token
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: Forbidden - Not allowed to perform this action
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Car not found
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      tags:
        - Cars
      summary: "Endpoint to delete a Car"
      description: "Deletes a Car. Requires token authentication, employee must be logged in to delete the ticket"
      consumes:
        - application/json
      produces:
        - application/json
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: "ID of the Car to delete"
      responses:
        200:
          description: "Car successfully deleted"
          schema:
            $ref: "#/definitions/DeleteCarResponse"
          examples:
            application/json:
              message: "Successfully deleted Car <car_id>"
        401:
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: Forbidden - Admin token required
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Car not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /cars/search:
    get:
      tags:
        - Cars
      summary: "Search cars by make"
      description: "Returns a list of cars that match the specified make. Supports partial matches (case-insensitive)"
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: make
          in: query
          required: true
          type: string
          description: "Make of the car to search for (e.g., Toyota, BMW)"
      responses:
        200:
          description: "Successfully retrieved matching cars."
          schema:
            type: array
            items:
              $ref: "#/definitions/CreateCarResponse"
          examples:
            application/json:
              - id: 3
                color: white
                make: Toyota
                model: Corolla
                model_year: 2021
                customer_id: 5
              - id: 7
                color: black
                make: Toyota
                model: Camry
                model_year: 2022
                customer_id: 2
        400:
          description: "Missing or invalid query parameter"
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ErrorResponse:
    type: "object"
    properties:
      message:
        type: "string"
...
  CreateCustomerCarPayload:
    type: "object"
    required:
      - color
      - make
      - model
      - model_year
      - customer_id
    properties:
      color:
        type: "string"
        description: "Color of the car customer is adding"
      make:
        type: "string"
        description: "The brand of the car (e.g. Kia, BMW, Toyota)"
      model:
        type: "string"
        description: "The model name of the car (e.g. Forte, M4, Camry)"
      model_year:
        type: "integer"
        description: "The year of the car when it was made"
      customer_id:
        type: "integer"
        description: "The ID of the customer you are adding to must be the same as the customer logged in"
...
  CreateCarPayload:
    type: object
    required:
      - color
      - make
      - model
      - model_year
      - customer_id
    properties:
      color:
        type: string
        description: "Color of the car"
      make:
        type: string
        description: "Make of the car"
      model:
        type: string
        description: "Model of the car"
      model_year:
        type: integer
        description: "Year the car was made"
      customer_id:
        type: integer
        description: "ID of the customer the car belongs to"
  CreateCarResponse:
    type: "object"
    properties:
      id:
        type: integer
      color:
        type: string
      make:
        type: string
      model:
        type: string
      model_year:
        type: integer
      customer_id:
        type: integer
  AllCars:
    type: "object"
    properties:
      cars:
        type: "array"
        items:
          $ref: "#/definitions/CreateCarResponse"
      page:
        type: "integer"
      per_page:
        type: "integer"
      total_cars:
        type: "integer"
  UpdateCarPayload:
    allOf:
      - $ref: "#/definitions/CreateCarResponse"
  UpdateCarResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      customer:
        $ref: "#/definitions/CreateCarResponse"
  DeleteCarResponse:
    type: "object"
    properties:
      message:
        type: "string"
...